chore(): refactor validation
